Remove dead compatibility code with ripper
Rails now requires Ruby 3.3, which always contains prism as a default gem
