+++ actionview/test/template/dependency_tracker_test.rb
@@ -3,11 +3,6 @@
 require "abstract_unit"
 require "action_view/dependency_tracker"
 
-require "action_view/render_parser/prism_render_parser"
-
-require "ripper"
-require "action_view/render_parser/ripper_render_parser"
-
 class NeckbeardTracker
   def self.call(name, template)
     ["foo/#{name}"]
@@ -293,9 +288,11 @@ class ERBTrackerTest < ActiveSupport::TestCase
   end
 end
 
-module RubyTrackerTests
+class RubyTrackerTest < Minitest::Test
+  include SharedTrackerTests
+
   def make_tracker(name, template, view_paths = nil)
-    ActionView::DependencyTracker::RubyTracker.new(name, template, view_paths, parser_class: parser_class)
+    ActionView::DependencyTracker::RubyTracker.new(name, template, view_paths)
   end
 
   def test_dependencies_skip_unknown_options
@@ -323,23 +320,5 @@ module RubyTrackerTests
     tracker = make_tracker("messages/show", template)
 
     assert_equal [], tracker.dependencies
-  end
-end
-
-class RipperRubyTrackerTest < ActiveSupport::TestCase
-  include SharedTrackerTests
-  include RubyTrackerTests
-
-  def parser_class
-    ActionView::RenderParser::RipperRenderParser
-  end
-end
-
-class PrismRubyTrackerTest < ActiveSupport::TestCase
-  include SharedTrackerTests
-  include RubyTrackerTests
-
-  def parser_class
-    ActionView::RenderParser::PrismRenderParser
   end
 end
